requirements-pip.txt
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2024.2.0
h5py==3.10.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pip==24.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
setuptools==68.2.2
six==1.16.0
sympy==1.12
threadpoolctl==3.3.0
tifffile==2024.2.12
torch==2.2.1
torchaudio==2.2.1
torchtext==0.17.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
